Refactor into separate graph module
