+++ trustchain-core/src/lib.rs
@@ -4,6 +4,7 @@ pub mod chain;
 pub mod controller;
 pub mod data;
 pub mod display;
+pub mod graph;
 pub mod key_manager;
 pub mod resolver;
 pub mod subject;
